requirements.txt
ipykernel
tensorflow==2.15.0
numpy
python-dotenv
pandas
matplotlib
seaborn
scikit-learn
tensorboard
flask
memory_profiler
streamlit
langchain
langchain-community
nltk
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
chromadb
jsonify
requests
sentence_transformers
langchain_huggingface
langchain_chroma
faiss-cpu
langchain_ollama
langchain-openai
langsmith
langchain_groq
fastapi
uvicorn
langserve